branch: mark and color a branch that is checked out in a linked worktree differently
In order to more clearly display which branches are active, modify
the output of git branch to mark branches checked out in a linked
worktree with a "+" and color them in cyan (in contrast to the
current branch, which will still be denoted with a "*" and colored
in green).

This is meant to simplify workflows related to worktree,
particularly due to the safety to refuse checking out the same
branch in two worktrees and deleting a branch checked out in a
worktree. When performing branch operations like checkout and
delete, it would be useful to know more readily if the branches in
which the user is interested are already checked out in another
worktree.

The git worktree list command contains the relevant information,
however this is a much less frquently used command than git branch.