ls-refs: filter refs based on non-namespace name
Protocol v2 allows for filtering of ref advertisement based on a
client-given name prefix, but inclusion/exclusion is determined based on
the non-namespace-stripped version (e.g. matching a ref prefix of
"refs/heads" against "refs/namespaces/my-namespace/refs/heads/master")
instead of the namespace-stripped version, which is the one that the
user actually sees.

Determine inclusion/exclusion based on the namespace-stripped version.

This bug was discovered through applying patches [1] that override
protocol.version to 2 in repositories when running tests, allowing us to
notice differences in behavior across different protocol versions.

[1] https://public-inbox.org/git/cover.1547677183.git.jonathantanmy@google.com/

Signed-off-by: Jonathan Tan <jonathantanmy@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>